+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 eSockd ChangeLog
 ==================
 
+2.5.0-beta (2015/05/01)
+------------------------
+
+support 'sockopts' option
+
+remove esockd:sockopts/1
+
+
 2.4.0-beta (2015/04/29)
 ------------------------